Code cleanup

+++ Api/0.0.0.1/EtAlii.Ubigia.Api.Functional.Scripting.NET47.SignalR.Tests/1. Logical/IntegrationTests/Win32HierarchicalContentManager.Tests.cs
@@ -25,10 +25,22 @@
 
         public void Dispose()
         {
-            if (Directory.Exists(_testFolderSimple))
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            // Cleanup
+            if (disposing && Directory.Exists(_testFolderSimple))
             {
                 Directory.Delete(_testFolderSimple, true);
             }
+        }
+
+        ~NET47HierarchicalContentManagerTests()
+        {
+            Dispose(false);
         }
 
         [Fact]
